\name{plot.vmapp}
\alias{plot.vmapp}
\title{
Plot Method for Class 'vmapp'
}
\description{
Plots a series of visualisations of vmapp.
}
\usage{
\method{plot}{vmapp}(x, ...)
}
\arguments{
\item{x}{
An object of class \code{"vmapp"}.
}
\item{...}{
additional arguments to be passed to the plot function.
}
}
\details{
This function is a method for the generic function
\code{plot()} for class \code{"vmapp"}.
It can be invoked by calling \code{plot(x)} for an
object \code{x} of the appropriate class, or directly by
calling \code{plot.vmapp(x)} regardless of the
class of the object.
}
\references{
    Chivers, C., Leung, B. Yann, N. (2013) Probabilistic predictions in ecology and their validation. In prep.
}
\author{
Corey Chivers <corey.chivers@mail.mcgill.ca>
}

