\name{print.vmapp}
\alias{print.vmapp}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Print Method for Class 'vmapp'
}
\description{
Prints a summary output of a vmapp object.
}
\usage{
print.vmapp(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
An object of type \code{"vmapp"}.
}
  \item{\dots}{
Additional arguments. These will be ignored.
}
}
\details{
This function is a method for the generic function
\code{print()} for class \code{"vmapp"}.
}
\value{
NULL
}
\references{
    Chivers, C., Leung, B. Yan, N. (2013) Probabilistic predictions in ecology and their validation. In prep.
}
\author{
Corey Chivers <corey.chivers@mail.mcgill.ca>
}

\seealso{
\code{\link{plot.vmapp}},\code{\link{vmapp}} 
}
