\name{rvmapp-package}
\alias{rvmapp-package}
\alias{rvmapp}
\docType{package}
\title{
Validation Metric Applied to Probabilistic Predictions
}
\description{
The Validation Metric Applied to Probabilistic Predictions (VMAPP) estimates the magnitude of differences between binary probabilistic model predictions and reality. VMAPP compares simulated outcomes from model predictions with the observed outcomes from a series of data not used to construct the model (ie. validation data) and determines if and where in the predictive range a model fails. 
}
\details{
\tabular{ll}{
Package: \tab rvmapp\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-06-05\cr
License: \tab GPL (>=3)\cr
}

}
\author{
Corey Chivers <corey.chivers@mail.mcgill.ca> & Brian Leung.
}
\references{
    Chivers, C., Leung, B. Yann, N. (2013) Probabilistic predictions in ecology and their validation. Ecology Letters.
}
